refactor: added exception throwing

+++ api-ecommerce/app/Repositories/ProductsRepository.ts
@@ -1,24 +1,31 @@
 import Product from "App/Models/Product"
 
 export default class ProductsRepository{
-  public async getAll(page?: number, limit?: number){
+  public async findAll(page?: number, limit?: number){
     const pagination = await Product.query().paginate(page || 1,limit || 5)
     return pagination.toJSON().data
   }
 
-  public async getById(productId: number){
-   return await Product.query().where('id',productId).select('*')
+  public async findById(productId: number){
+    const product =  await Product.findBy('id',productId)
+    if(!product) throw new Error('NOT_FOUND')
+    return product
   }
 
   public async store(product: Product){
-    return await Product.create(product)
+    await Product.create(product)
+    return {message: `Produto ${product.name} adicionado com sucesso!`}
   }
 
   public async update(productId : number, product: Product){
-    return await Product.query().where('id',productId).update(product)
+    if(!await this.findById(productId)) throw new Error('NOT_FOUND')
+    await Product.query().where('id',productId).update(product)
+    return {message: `Produto atualizado com sucesso!`}
   }
 
   public async delete(productId: number){
-    return await Product.query().where('id',productId).delete()
+    if(!await this.findById(productId)) throw new Error('NOT_FOUND')
+    await Product.query().where('id',productId).delete()
+    return {message: `Produto deletado com sucesso!`}
   }
 }
